+++ src/loader/parsers/video.js
@@ -2,18 +2,18 @@ import { videoList } from "../cache.js";
 import { fetchData } from "./fetchdata.js";
 import { hasVideoFormat } from "../../system/device.js";
 import * as fileUtil from "../../utils/file.js";
-import { crossOrigin } from "../settings.js";
 import { isDataUrl } from "./../../utils/string.js";
 
 /**
  * parse/preload a Video file
  * @param {loader.Asset} data - asset data
- * @param {Function} [onload] - function to be called when the asset is loaded
+ * @param {Function} [onload] - function to be called when the resource is loaded
  * @param {Function} [onerror] - function to be called in case of error
+ * @param {Object} [settings] - Additional settings to be passed when loading the asset
  * @returns {number} the amount of corresponding resource parsed/preloaded
  * @ignore
  */
-export function preloadVideo(data, onload, onerror) {
+export function preloadVideo(data, onload, onerror, settings) {
 
     if (typeof videoList[data.name] !== "undefined") {
         // Video already preloaded
@@ -34,7 +34,7 @@ export function preloadVideo(data, onload, onerror) {
     }
 
     if (isDataUrl(data.src)) {
-        fetchData(data.src, "blob")
+        fetchData(data.src, "blob", settings)
             .then(blob => {
                 videoElement.src = globalThis.URL.createObjectURL(blob);
             })
@@ -52,7 +52,7 @@ export function preloadVideo(data, onload, onerror) {
     videoElement.setAttribute("playsinline", "true");
     videoElement.setAttribute("disablePictureInPicture", "true");
     videoElement.setAttribute("controls", "false");
-    videoElement.setAttribute("crossorigin", crossOrigin);
+    videoElement.setAttribute("crossorigin", settings.crossOrigin);
 
     if (data.autoplay === true) {
         videoElement.setAttribute("autoplay", "true");