+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 ### Added
 - Sprite: add support for aseprite texture atlas  
 - Atlas: `createSpritefromAnim` parameter is now optional, and if not defined will use all defined index in the corresponding atlas
+- Loader: new `setOptions` method that allows specifying custom settings to be applied to fetch requests (crossOrigin, withCredentials, etc..)
 
 ### Fixed
 - Doc: fix hyperlinks to source code within documentation (thanks @Waltibaba)
+- Loader: fix settings for the fetch request not being applied in es6 builds (thanks @B0sh)
+
 
 ## [16.1.2] (melonJS 2) - _2024-02-12_